+++ DevTools/CharacterListProcessing.cs
@@ -12,7 +12,7 @@ namespace DevTools
 {
 	class CharacterListProcessing
 	{
-		public const string kBaseDirForRealOutput = "..\\..\\Glyssen\\Resources";
+		public const string kBaseDirForRealOutput = "..\\..\\GlyssenCharacters\\Resources";
 		private const string kBaseDirForHelperOutput = "..\\..\\Resources\\temporary";
 
 		public static void Process()